Crate xgrammar

Re-exports

pub use tokenizers;

Modules

huggingface_hub

Structs

CompiledGrammar

FromPretrainedParameters

Defines the additional parameters available for the from_pretrained function

Grammar

GrammarCompiler

GrammarMatcher

MetadataFromHF

StructuralTagItem

Represents a structural tag item with begin, schema, and end components. This is used for structured text generation with specific formatting tags.

TokenizerInfo

Enums

VocabType

Statics

HF_CONFIG_EOS_TOKEN_ID_KEY

TOKENIZER_ALLOW_PATTERN

TOKENIZER_CONFIG_FILE

TOKENIZER_FILE

TOKENIZER_MODEL_KEY

TOKENIZER_VOCAB_KEY

Type Aliases

TokenId

VocabMap