xgrammar

Module huggingface_hub

Structs

Api

The actual Api used to interacto with the hub. Use any repo with Api::repo

ApiBuilder

Helper to create Api with all the options.

ApiRepo

Shorthand for accessing things within a particular repo You can inspect repos with ApiRepo::info or download files with ApiRepo::download

Params

Repo

The representation of a repo on the hub.

RepoInfo

The description of the repo given by the hub

Enums

ApiError

All errors the API can throw

HuggingfaceError

RepoType

The type of repo to interact with

Functions

compile_glob_pattern

snapshot_download